==> 00_download_raw_data.ipynb <==
%reload_ext autoreload
%autoreload 2

import time
from src.data_green import load_green_raw_data

load_green_raw_data(year=2018)
time.sleep(30)

load_green_raw_data(year=2017)
time.sleep(30)

load_green_raw_data(year=2016)
time.sleep(30)

load_green_raw_data(year=2015)
time.sleep(30)

load_green_raw_data(year=2014)
time.sleep(30)
