==> 00_upload_raw_data_adls.ipynb <==
%reload_ext autoreload
%autoreload 2

from src.adls import upload_files_to_directory_bulk
import time

upload_files_to_directory_bulk("2018")
time.sleep(30)

upload_files_to_directory_bulk("2017")
time.sleep(30)

upload_files_to_directory_bulk("2016")
time.sleep(30)

upload_files_to_directory_bulk("2015")
time.sleep(30)

upload_files_to_directory_bulk("2014")
