==> 01_load_and_validate_raw_data.ipynb <==
!wget https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2022-01.parquet -O data/raw/rides_2022_01.parquet

import pandas as pd

rides = pd.read_parquet('data/raw/rides_2022_01.parquet')

rides.head(10)


rides = rides[['tpep_pickup_datetime', 'PULocationID']]

rides.rename(columns={
    'tpep_pickup_datetime': 'pickup_datetime',
    'PULocationID': 'pickup_location_id',
}, inplace=True)

rides.head(10)

rides['pickup_datetime'].describe()

rides = rides[rides.pickup_datetime >= '2022-01-01']
rides = rides[rides.pickup_datetime < '2022-02-01']
rides['pickup_datetime'].describe()

rides.to_parquet('data/transformed/validated_rides_2022_01.parquet')
